Include less/Less.php, not less/Cache.php
If the Cache module fails for some reason, it will attempt to
throw a Less_Exception_Parser type of exception.  However, if
only the less/Cache.php file is included, that class will not
have been defined.  (It is defined in "less/Less.php".)

So, include less/Less.php, so the Less_Exception_Parser class
can be defined, so when an error occurs, it can be reported
successfully.