FREEPBX-13317 - Handle errors in Password Reset
This moves the error to be an alert(), as if it's in the
login box it moves the button down too far.  See
http://i.imgur.com/13i8OeK.png

It also will ALWAYS send a password reset email, even
if one has already been sent, in case it was lost in
transit or eaten by a spam filter.